[refactor]: split components
